+++ src/crypto/conf/conf.c
@@ -222,7 +222,7 @@ static int str_copy(CONF *conf, char *section, char **pto, char *from) {
     return 0;
   }
 
-  len = strlen(from) + 1;
+  len = (int)(strlen(from) + 1);
   if (!BUF_MEM_grow(buf, len)) {
     goto err;
   }
@@ -572,7 +572,7 @@ static int def_load_bio(CONF *conf, BIO *in, long *out_error_line) {
     *p = '\0';
     BIO_gets(in, p, CONFBUFSIZE - 1);
     p[CONFBUFSIZE - 1] = '\0';
-    ii = i = strlen(p);
+    ii = i = (int)strlen(p);
     if (i == 0 && !again) {
       break;
     }
@@ -797,7 +797,7 @@ int CONF_parse_list(const char *list, char sep, int remove_whitespace,
           tmpend--;
         }
       }
-      ret = list_cb(lstart, tmpend - lstart + 1, arg);
+      ret = list_cb(lstart, (int)(tmpend - lstart + 1), arg);
     }
     if (ret <= 0) {
       return ret;
