+++ src/crypto/fipsmodule/bn/exponentiation.c
@@ -735,7 +735,7 @@ void bn_mod_exp_mont_small(BN_ULONG *r, const BN_ULONG *a, size_t num,
     bn_from_montgomery_small(r, num, mont->RR.d, num, mont);
     return;
   }
-  unsigned bits = BN_num_bits_word(p[num_p - 1]) + (num_p - 1) * BN_BITS2;
+  unsigned bits = (unsigned)(BN_num_bits_word(p[num_p - 1]) + (num_p - 1) * BN_BITS2);
   assert(bits != 0);
 
   // We exponentiate by looking at sliding windows of the exponent and
