+++ src/crypto/x509/a_sign.c
@@ -113,7 +113,7 @@ int ASN1_item_sign_ctx(const ASN1_ITEM *it,
         OPENSSL_free(signature->data);
     signature->data = buf_out;
     buf_out = NULL;
-    signature->length = outl;
+    signature->length = (int)outl;
     /*
      * In the interests of compatibility, I'll make sure that the bit string
      * has a 'not-used bits' value of 0
@@ -124,5 +124,5 @@ int ASN1_item_sign_ctx(const ASN1_ITEM *it,
     EVP_MD_CTX_cleanup(ctx);
     OPENSSL_free(buf_in);
     OPENSSL_free(buf_out);
-    return (outl);
+    return (int)(outl);
 }
