+++ src/crypto/x509/x509cset.c
@@ -124,7 +124,7 @@ int X509_CRL_sort(X509_CRL *c)
     sk_X509_REVOKED_sort(c->crl->revoked);
     for (i = 0; i < sk_X509_REVOKED_num(c->crl->revoked); i++) {
         r = sk_X509_REVOKED_value(c->crl->revoked, i);
-        r->sequence = i;
+        r->sequence = (int)i;
     }
     c->crl->enc.modified = 1;
     return 1;
@@ -275,7 +275,7 @@ int X509_CRL_set1_signature_algo(X509_CRL *crl, const X509_ALGOR *algo)
 int X509_CRL_set1_signature_value(X509_CRL *crl, const uint8_t *sig,
                                   size_t sig_len)
 {
-    if (!ASN1_STRING_set(crl->signature, sig, sig_len)) {
+    if (!ASN1_STRING_set(crl->signature, sig, (int)sig_len)) {
       return 0;
     }
     crl->signature->flags &= ~(ASN1_STRING_FLAG_BITS_LEFT | 0x07);
